requirements.txt
discord.py==2.4.0
discord_webhook==1.3.1
pymongo==4.6.1
python-dotenv==1.0.1
